clean up julia code a bit
